Обновил ветку на девелопе новой версией
