+++ .gitignore
@@ -1,0 +1,4 @@
+/src
+.idea
+/node_modules
+/dist
