+++ package.json
@@ -1,14 +1,20 @@
 {
   "name": "supersborka",
-  "version": "1.0.0",
+  "version": "0.0.6",
   "description": "SdelauSuperLightSborku",
   "main": "script.js",
   "scripts": {
     "build": "cross-env NODE_ENV=production  rimraf dist && webpack --mode production",
-    "dev": "cross-env NODE_ENV=development webpack-dev-server --mode development --open --watch"
+    "dev": "cross-env NODE_ENV=development webpack-dev-server --mode development --open --watch",
+    "PushMe": "cross-env NODE_ENV=production gh-pages -d dist"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/VitalyBibik/pushme"
   },
   "author": "VitalyBibik",
   "license": "ISC",
+  "homepage": "https://github.com/VitalyBibik/pushme/",
   "devDependencies": {
     "@babel/cli": "^7.8.4",
     "@babel/core": "^7.9.0",
@@ -20,6 +26,7 @@
     "css-loader": "^3.5.2",
     "cssnano": "^4.1.10",
     "file-loader": "^6.0.0",
+    "gh-pages": "~2.0.1",
     "html-webpack-plugin": "^3.2.0",
     "image-webpack-loader": "^6.0.0",
     "mini-css-extract-plugin": "^0.9.0",
