+++ postcss.config.js
@@ -6,3 +6,4 @@ module.exports = {
         })
     ]
 }
+// 65 еслпи писать browser:last 10 version
