+++ src/index.html
@@ -92,10 +92,8 @@
   </div>
 
 
- <!--<script src="./script/initialCard.js"></script> -->
+
 <script src="<%= htmlWebpackPlugin.files.chunks.main.entry %>"></script>
 
-
-  <!-- классы -> константы + initialsCards + инициализация классов-->
 </body>
 </html>
