+++ src/scripts/UserInfo.js
@@ -1,15 +1,3 @@
-/* Класс для работы с данными пользователя. Экземпляр этого класса должен хранить в себе
- данные пользователя: имя и информацию о себе, а также отображать эту информацию на странице.
-Для этого класса нужно определить методы:
-
-setUserInfo, чтобы обновлять данные внутри экземпляра класса;
-updateUserInfo, чтобы отображать эти данные на странице.
-
-Логичный вопрос: почему не объединить эти методы в один, который бы обновлял данные и затем выводил на экран?
-Первая причина — принцип разделения ответственности. Лучше, чтобы каждый метод отвечал за небольшую часть функциональности.
-Вторая причина станет понятна в следующем спринте, — когда мы подключим проект к серверу. Тогда, чтобы обновить данные,
- сначала нужно будет отправить запрос на сервер, дождаться ответа и только после этого обновить DOM.
- Поэтому лучше сразу вынести обновление DOM в отдельную функцию.*/
 
 class UserInfo {
   constructor(name, job, nameCont, jobCont, api, spinner) {
@@ -48,10 +36,6 @@ class UserInfo {
 
     const user = this.api.setUser(this.nameCont.textContent, this.jobCont.textContent)
         .catch(error => console.log(error));
-     /*    .finally(() => {
-      this.renderLoading(false);
-    });*/
-    // Отправляю данные на сервер,
 
     this.namepage = this.nameCont.textContent;
     this.jobpage = this.jobCont.textContent;
