+++ src/scripts/script.js
@@ -141,36 +141,3 @@ placesList.addEventListener('click', popupZoom.open.bind(popupZoom));
 buttonAvatar.addEventListener('click', popupAvatar.open.bind(popupAvatar));
 buttonAvatarSave.addEventListener('click', popupAvatar.sendAvatar.bind(popupAvatar));
 
-
-/**
-* Здравствуйте.
-*
-* Очень хорошая работа.
- Но хотелось бы чтобы вы поправили
- Данные об адресе "https://praktikum.tk/cohort9/" использовали. А то вы передаёте в baseUrl, но не используете
-    то же данные по авторизации '41b0685a-8626-46fa-882b-88da0ea48249'
-    Надо исправить: Необходимо в методы добавить обработку ошибок
-      .catch((err) => {
-    	console.log(err);
-    	});
-
-    * Класс Api это отдельный класс который ничего не знает о других классах и методах
-    * Вы можете только получать данные из этого класса и использовать эти данные.
-    * Представьте, что я дам Вам другой класс(допустим DataBase) к внутренностям которого вы не будете иметь доступ и даже прочитать этот файл не сможете
-    * скажу что у него есть несколько методов  getInitialCards deleteCard addCard, editUserInfo, setUserInfo и так далее
-    * Который только возвращает данные, а вы можите получить только обращась к этим методам.
-    * Соответственно в классе нельзя реализовать такие методы как querySelector или обратиться к другому классу, а только обратиться к методам.
-    * Отдельная обязанность. Таким же способом Вы обращаетесь к серверу. Вы не знаете, что на сервере, даже язык программирования, но вы знаете методы
-    * к которым обращаетесь и способ обращения. Это и есть обязанность отдельного класса.
-    *
-
-*
- *
-*/
-
-/**
- * Здравствуйте,
- * работа принимается
- *
- *
- */
