+++ webpack.config.js
@@ -26,8 +26,15 @@ module.exports = {
                 test: /\.css$/i,
                 use: [
                     (isDev ? 'style-loader' : MiniCssExtractPlugin.loader),
-                    'css-loader',
-                    'postcss-loader'
+                    {
+                        loader:'css-loader',
+                        options: {
+                            importLoaders: 2
+                        }
+                    },
+                    {
+                        loader: "postcss-loader",
+                    }
                 ]
             },
             {
